Split anonymous free-variable analysis
- Move anonymous function free-variable collection into a dedicated child module.
- Keep function.rs focused on anonymous function lowering and closure construction.
- Colocate free-variable collector tests with the module that owns the traversal.
- Preserve existing planner behavior and anonymous function plan shapes.
